pyproject.toml
[project]
name = "awscli"
version = "0.1.0"
description = ""
authors = [
    {name = "SMARTSOFT\\a.esakia",email = "a.esakia@smartsoft.ge"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all] (>=0.15.2,<0.16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "boto3 (>=1.37.18,<2.0.0)",
    "python-magic-bin (>=0.4.14,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
awsclitool = "awscli.main:app"

